refactor: criação de um novo folder para melhor visualização das injeções de dependência
